Update CI config
When deploying the built documentation files using the action by
crazy-max, it would complain[1] that it doesn't want to copy the
'result' symlink we get from running 'nix build' to the folder where
the deployed stuff is. That's why we need to export the actual store
location into an output and use that as the build directory. Further,
the action expects a relative path[2], so we need to traverse the
filesystem back up all the way so we can get to the nix store.

1: https://github.com/iWeltAG/zucker/runs/4427892335
2: https://github.com/crazy-max/ghaction-github-pages/issues/159