Use SyncModule and AsyncModule directly
This removes the `client.Module` syntax we had before, because that was
a nightmare to type correctly. This is more or less due to this[1] issue
in mypy. Once we find a good way to make type checkers happy, the syntax
may return. But for now, using SyncModule and AsyncModule as a
superclass will suffice.

1: https://github.com/python/mypy/issues/10962