Evaluate view slices lazily
These changes make sure that we can correctly build ranges for our
views. The problem was that evaulation the view range may depend on
knowing the exact number of records the server has. This is a problem
whenever we have a slice that is counting "from the back", for example:

  [-4:]
  [1:-5]
  [-10:20:2]

In that case, we need to know the total number in order to perform the
counting. The way this is solved now is by queueing up all slicing
operations and only evaluating them when we are in a position where we
will be performing requests anyway (for async views, this is will be
inside an asynchronous call).