Refactor request handling int base client classes
This is in preparation for #7, allowing bulk request handling. In order
to make this more maintainable, we want to lift the burden on the
individual client implementations and handle everything in the base
class. To do this, we introduce a generic raw_request() method that
does the actual server communication and will be overridden by
implementations. request(), which is the one that should be called,
is now implemented in the sync and async superclasses.