Add basic async iteration implementation
This implementation should more or less work, although it isn't perfect
and could still be improved by properly fetching ranges of records
instead of stitching together individual indexes using the /bulk
endpoint. It does, however, skip fetching records that are not in range
if the view does something like [::2] with a skip option - so at least
it's got that going for it.

Closes #1.