Fix live_async_client test fixture
Before the fixture wouldn't actually return the client, but rather the
asynchronous generator (which we don't actually need). To fix this, the
fixture is now synchronous again and closes the client by opening an
event loop again when finalizing (because Pytest's event loop is already
closed by then, hence we need a new one).